requirements.txt
Flask
Flask-Login==0.4.0
Flask-WTF==0.13.0
Jinja2==2.8
MarkupSafe==0.23
WTForms==2.1
Werkzeug==0.10.4
argparse==1.2.1
flask-mongoengine==0.8.2
itsdangerous==0.24
mongoengine==0.11.0
pymongo==3.4.0
request==2.2.1
six==1.10.0
wsgiref==0.1.2